mdms/mdms/Backend/requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.0.0
exceptiongroup==1.3.1
fastapi==0.95.2
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Pillow==10.0.0
imagehash>=4.3.1
psycopg2-binary==2.9.6
pydantic==1.10.26
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.22.0
watchfiles==1.1.1
websockets==15.0.1

# YOLOv5 Dependencies
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.1.1
numpy>=1.23.5
ultralytics>=8.2.64
matplotlib>=3.3
scipy>=1.4.1
pandas>=1.1.4
seaborn>=0.11.0
tqdm>=4.66.3
gitpython>=3.1.30
psutil
packaging
setuptools>=70.0.0